Move creating locations and services points to API test base class CIRC-184
